adapted my solution for part 2 from Hallski's solution
I poured enough time into this and I just want to be done. This is
similar to how my code was structured except I was passing the results
back instead of appending the rules to the front of a queue. I thought
about something like that, but was too burnt to actually implement it.

Either way, it's a good lesson. You can't back track when you are
returning the result. That function invocation has thrown it's other
options away.

Instead you need to keep all of the remaining state available going
forward.

Link to post here:

https://elixirforum.com/t/avent-of-code-2020-day-19/36324/8